+++ internal/core/window.rs
@@ -597,6 +597,7 @@ pub struct WindowInner {
     /// ItemRC that currently have the focus (possibly an instance of TextInput)
     pub focus_item: RefCell<crate::item_tree::ItemWeak>,
     focus_item_visibility_tracker: ChangeTracker,
+    focus_item_position_tracker: ChangeTracker,
     /// The last text that was sent to the input method
     pub(crate) last_ime_text: RefCell<SharedString>,
     /// Don't let ComponentContainers's instantiation change the focus.
@@ -675,6 +676,7 @@ impl WindowInner {
             }),
             focus_item: Default::default(),
             focus_item_visibility_tracker: Default::default(),
+            focus_item_position_tracker: Default::default(),
             last_ime_text: Default::default(),
             cursor_blinker: Default::default(),
             active_popups: Default::default(),
@@ -694,6 +696,7 @@ impl WindowInner {
     pub fn set_component(&self, component: &ItemTreeRc) {
         self.close_all_popups();
         self.focus_item_visibility_tracker.clear();
+        self.focus_item_position_tracker.clear();
         self.focus_item.replace(Default::default());
         self.mouse_input_state.replace(Default::default());
         self.touch_state.replace(Default::default());
@@ -1435,6 +1438,7 @@ impl WindowInner {
     /// This sends the event which must be either FocusOut or WindowLostFocus for popups
     fn take_focus_item(&self, event: &FocusEvent) -> Option<ItemRc> {
         self.focus_item_visibility_tracker.clear();
+        self.focus_item_position_tracker.clear();
         let focus_item = self.focus_item.take();
         assert!(matches!(event, FocusEvent::FocusOut(_)));
 
@@ -1468,7 +1472,7 @@ impl WindowInner {
                 );
                 // Reveal offscreen item when it gains focus
                 if result == crate::input::FocusEventResult::FocusAccepted {
-                    self.track_focus_item_visibility(item);
+                    self.track_focus_item(item);
                     item.try_scroll_into_visible();
                 }
 
@@ -1476,13 +1480,15 @@ impl WindowInner {
             }
             None => {
                 self.focus_item_visibility_tracker.clear();
+                self.focus_item_position_tracker.clear();
                 *self.focus_item.borrow_mut() = Default::default();
                 crate::input::FocusEventResult::FocusAccepted // We were removing the focus, treat that as OK
             }
         }
     }
 
-    fn track_focus_item_visibility(&self, item: &ItemRc) {
+    fn track_focus_item(&self, item: &ItemRc) {
+        // Track visibility
         let visibility_clips = item.visibility_clips();
         self.focus_item_visibility_tracker.init(
             (item.downgrade(), self.window_adapter_weak.clone(), visibility_clips),
@@ -1502,6 +1508,24 @@ impl WindowInner {
                     false,
                     FocusReason::Programmatic,
                 );
+            },
+        );
+
+        // Track position
+        self.focus_item_position_tracker.init(
+            (item.downgrade(), self.window_adapter_weak.clone()),
+            |(item, _)| {
+                let Some(item) = item.upgrade() else { return Default::default() };
+                Some(item.map_to_native_window(item.geometry().origin))
+            },
+            |(item, window_adapter), _| {
+                let (Some(item), Some(window_adapter)) = (item.upgrade(), window_adapter.upgrade())
+                else {
+                    return;
+                };
+                if let Some(text_input) = item.downcast::<crate::items::TextInput>() {
+                    text_input.as_pin_ref().update_ime(&window_adapter, &item);
+                }
             },
         );
     }